Migrate BrigAccess to wire-subsystems
