+++ libs/wire-subsystems/wire-subsystems.cabal
@@ -179,6 +179,8 @@ library
     Wire.AWS
     Wire.BlockListStore
     Wire.BlockListStore.Cassandra
+    Wire.BrigAPIAccess
+    Wire.BrigAPIAccess.Rpc
     Wire.DeleteQueue
     Wire.DeleteQueue.InMemory
     Wire.DomainRegistrationStore
@@ -312,6 +314,7 @@ library
     , HsOpenSSL
     , hspec
     , html-entities
+    , http-api-data
     , http-client
     , http-types
     , http2-manager
@@ -379,6 +382,7 @@ test-suite wire-subsystems-tests
     Spec
     Wire.ActivationCodeStore.InterpreterSpec
     Wire.AuthenticationSubsystem.InterpreterSpec
+    Wire.BrigAPIAccess.RpcSpec
     Wire.EnterpriseLoginSubsystem.InterpreterSpec
     Wire.HashPassword.InterpreterSpec
     Wire.MiniBackend
