+++ services/galley/src/Galley/API/Create.hs
@@ -52,7 +52,6 @@ import Galley.API.Teams.Features.Get (getFeatureForTeam)
 import Galley.API.Util
 import Galley.App (Env)
 import Galley.Effects
-import Galley.Effects.BrigAccess
 import Galley.Effects.ConversationStore qualified as E
 import Galley.Effects.FederatorAccess qualified as E
 import Galley.Effects.MemberStore qualified as E
@@ -84,6 +83,7 @@ import Wire.API.Team.LegalHold (LegalholdProtectee (LegalholdPlusFederationNotIm
 import Wire.API.Team.Member
 import Wire.API.Team.Permission hiding (self)
 import Wire.API.User
+import Wire.BrigAPIAccess
 import Wire.NotificationSubsystem
 import Wire.Sem.Now (Now)
 import Wire.Sem.Now qualified as Now
@@ -99,7 +99,7 @@ import Wire.UserList
 -- API up to and including version 3.
 createGroupConversationUpToV3 ::
   ( Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (Error FederationError) r,
@@ -143,7 +143,7 @@ createGroupConversationUpToV3 lusr conn newConv = mapError UnreachableBackendsLe
 -- API in from version 4 to 8
 createGroupOwnConversation ::
   ( Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (Error FederationError) r,
@@ -188,7 +188,7 @@ createGroupOwnConversation lusr conn newConv = do
 -- API in version 9 and above.
 createGroupConversation ::
   ( Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (Error FederationError) r,
@@ -253,7 +253,7 @@ createGroupConvAndMkResponse ::
     Member P.TinyLog r,
     Member FederatorAccess r,
     Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member NotificationSubsystem r,
     Member LegalHoldStore r,
@@ -276,7 +276,7 @@ createGroupConvAndMkResponse lusr conn newConv mkResponse = do
 createGroupConversationGeneric ::
   forall r.
   ( Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (Error FederationError) r,
@@ -355,7 +355,7 @@ ensureNoLegalholdConflicts (UserList locals remotes) = do
       throwS @'MissingLegalholdConsent
 
 checkCreateConvPermissions ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (ErrorS 'NotATeamMember) r,
     Member (ErrorS OperationDenied) r,
@@ -472,7 +472,7 @@ createProteusSelfConversation lusr = do
 
 createOne2OneConversation ::
   ( Member BackendNotificationQueueAccess r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ConversationStore r,
     Member (Error FederationError) r,
     Member (Error InternalError) r,
