+++ services/galley/src/Galley/API/MLS/Reset.hs
@@ -60,7 +60,7 @@ resetMLSConversation ::
     Member FederatorAccess r,
     Member ExternalAccess r,
     Member (Error FederationError) r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member MemberStore r,
     Member NotificationSubsystem r,
     Member ProposalStore r,
@@ -108,7 +108,7 @@ resetRemoteMLSConversation ::
     Member (ErrorS MLSStaleMessage) r,
     Member (Error FederationError) r,
     Member (Error InternalError) r,
-    Member BrigAccess r,
+    Member BrigAPIAccess r,
     Member ExternalAccess r,
     Member FederatorAccess r,
     Member NotificationSubsystem r,
