+++ services/galley/src/Galley/API/Util.hs
@@ -47,7 +47,6 @@ import Galley.Data.Services (BotMember, newBotMember)
 import Galley.Data.Types qualified as DataTypes
 import Galley.Effects
 import Galley.Effects.BackendNotificationQueueAccess
-import Galley.Effects.BrigAccess
 import Galley.Effects.ClientStore
 import Galley.Effects.CodeStore
 import Galley.Effects.ConversationStore
@@ -93,6 +92,7 @@ import Wire.API.Team.Role
 import Wire.API.User hiding (userId)
 import Wire.API.User.Auth.ReAuth
 import Wire.API.VersionInfo
+import Wire.BrigAPIAccess
 import Wire.HashPassword (HashPassword)
 import Wire.HashPassword qualified as HashPassword
 import Wire.NotificationSubsystem
@@ -106,7 +106,7 @@ import Wire.UserList
 data NoChanges = NoChanges
 
 ensureAccessRole ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'NotATeamMember) r,
     Member (ErrorS 'ConvAccessDenied) r
   ) =>
@@ -128,7 +128,7 @@ ensureAccessRole roles users = do
 -- | Check that the given user is either part of the same team as the other
 -- users OR that there is a connection.
 ensureConnectedOrSameTeam ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'NotConnected) r,
     Member TeamStore r,
     Member TeamCollaboratorsSubsystem r
@@ -149,7 +149,7 @@ ensureConnectedOrSameTeam lusr others = do
 -- 'ensureConnected' for non-team-members of the _given_ user. Implicit
 -- connections are created per team, so we count them as team membership here.
 ensureConnectedToLocalsOrSameTeam ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'NotConnected) r,
     Member TeamStore r,
     Member TeamCollaboratorsSubsystem r
@@ -178,7 +178,7 @@ ensureConnectedToLocalsOrSameTeam (tUnqualified -> u) uids = do
 -- B blocks A, the status of A-to-B is still 'Accepted' but it doesn't mean
 -- that they are connected).
 ensureConnected ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'NotConnected) r
   ) =>
   Local UserId ->
@@ -190,7 +190,7 @@ ensureConnected self others = do
 
 ensureConnectedToLocals ::
   ( Member (ErrorS 'NotConnected) r,
-    Member BrigAccess r
+    Member BrigAPIAccess r
   ) =>
   UserId ->
   [UserId] ->
@@ -203,7 +203,7 @@ ensureConnectedToLocals u uids = do
     throwS @'NotConnected
 
 ensureConnectedToRemotes ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'NotConnected) r
   ) =>
   Local UserId ->
@@ -216,7 +216,7 @@ ensureConnectedToRemotes u remotes = do
     throwS @'NotConnected
 
 ensureReAuthorised ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (Error AuthenticationError) r
   ) =>
   UserId ->
@@ -797,7 +797,7 @@ verifyReusableCode rateLimitKey checkPw mPtpw convCode = do
   pure c
 
 ensureConversationAccess ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member (ErrorS 'ConvAccessDenied) r,
     Member (ErrorS 'NotATeamMember) r,
     Member TeamStore r
@@ -1177,7 +1177,7 @@ getLocalUsers :: Domain -> NonEmpty (Qualified UserId) -> [UserId]
 getLocalUsers localDomain = map qUnqualified . filter ((== localDomain) . qDomain) . toList
 
 getBrigClients ::
-  ( Member BrigAccess r,
+  ( Member BrigAPIAccess r,
     Member ClientStore r
   ) =>
   [UserId] ->
