--- services/galley/src/Galley/External/LegalHoldService/Types.hs
@@ -1,40 +1,0 @@
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-
--- This file is part of the Wire Server implementation.
---
--- Copyright (C) 2022 Wire Swiss GmbH <opensource@wire.com>
---
--- This program is free software: you can redistribute it and/or modify it under
--- the terms of the GNU Affero General Public License as published by the Free
--- Software Foundation, either version 3 of the License, or (at your option) any
--- later version.
---
--- This program is distributed in the hope that it will be useful, but WITHOUT
--- ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
--- FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
--- details.
---
--- You should have received a copy of the GNU Affero General Public License along
--- with this program. If not, see <https://www.gnu.org/licenses/>.
-
-module Galley.External.LegalHoldService.Types
-  ( OpaqueAuthToken (..),
-
-    -- * Re-exports
-    LegalHoldService,
-  )
-where
-
-import Brig.Types.Team.LegalHold
-import Data.Aeson
-import Data.ByteString.Conversion.To
-import Imports
-
--- | When receiving tokens from other services which are 'just passing through'
--- it's error-prone useless extra work to parse and render them from JSON over and over again.
--- We'll just wrap them with this to give some level of typesafety and a reasonable JSON
--- instance
-newtype OpaqueAuthToken = OpaqueAuthToken
-  { opaqueAuthTokenToText :: Text
-  }
-  deriving newtype (Eq, Show, FromJSON, ToJSON, ToByteString)
