+++ services/galley/test/unit/Run.hs
@@ -24,7 +24,6 @@ import Imports
 import Test.Galley.API.Action qualified
 import Test.Galley.API.Message qualified
 import Test.Galley.API.One2One qualified
-import Test.Galley.Intra.User qualified
 import Test.Galley.Mapping qualified
 import Test.Tasty
 
@@ -35,7 +34,6 @@ main =
       "Tests"
       [ Test.Galley.API.Message.tests,
         Test.Galley.API.One2One.tests,
-        Test.Galley.Intra.User.tests,
         Test.Galley.Mapping.tests,
         Test.Galley.API.Action.tests
       ]
