--- services/galley/test/unit/Test/Galley/Intra/User.hs
@@ -1,44 +1,0 @@
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ScopedTypeVariables #-}
--- Disabling to stop warnings on HasCallStack
-{-# OPTIONS_GHC -Wno-redundant-constraints #-}
-
--- This file is part of the Wire Server implementation.
---
--- Copyright (C) 2022 Wire Swiss GmbH <opensource@wire.com>
---
--- This program is free software: you can redistribute it and/or modify it under
--- the terms of the GNU Affero General Public License as published by the Free
--- Software Foundation, either version 3 of the License, or (at your option) any
--- later version.
---
--- This program is distributed in the hope that it will be useful, but WITHOUT
--- ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
--- FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
--- details.
---
--- You should have received a copy of the GNU Affero General Public License along
--- with this program. If not, see <https://www.gnu.org/licenses/>.
-
-module Test.Galley.Intra.User where
-
-import Galley.Intra.User (chunkify)
-import Imports
-import Test.QuickCheck
-import Test.Tasty
-import Test.Tasty.QuickCheck
-
-tests :: TestTree
-tests =
-  testGroup
-    "Tests"
-    [ testChunkify
-    ]
-
-testChunkify :: (HasCallStack) => TestTree
-testChunkify =
-  testGroup
-    "chunkify"
-    [ testProperty "is inverse of chunk concatenation" . mapSize (* 314) $ \(xs :: [Int]) ->
-        chunkify pure xs === Just xs
-    ]
